+++ lib/gui/app/i18n/en.ts
@@ -10,7 +10,7 @@ const translation = {
 		failed: 'Failed',
 		completed: 'Completed',
 		yesContinue: 'Yes, continue',
-		reallyExit: 'Are you sure you want to close Etcher?',
+		reallyExit: 'Are you sure you want to close Spark?',
 		yesExit: 'Yes, quit',
 		progress: {
 			starting: 'Starting...',
@@ -39,7 +39,7 @@ const translation = {
 			andFailTarget_other: 'and failed to be flashed to {{num}} targets',
 			succeedTo: '{{name}} was successfully flashed {{target}}',
 			exitWhileFlashing:
-				'You are currently flashing a drive. Closing Etcher may leave your drive in an unusable state.',
+				'You are currently flashing a drive. Closing Spark may leave your drive in an unusable state.',
 			looksLikeWindowsImage:
 				'It looks like you are trying to burn a Windows image.\n\nUnlike other images, Windows images require special processing to be made bootable. We suggest you use a tool specially designed for this purpose, such as <a href="https://rufus.akeo.ie">Rufus</a> (Windows), <a href="https://github.com/slacka/WoeUSB">WoeUSB</a> (Linux), or Boot Camp Assistant (macOS).',
 			image: 'image',
@@ -56,16 +56,16 @@ const translation = {
 			genericFlashError:
 				'Something went wrong. If it is a compressed image, please check that the archive is not corrupted.\n{{error}}',
 			validation:
-				'The write has been completed successfully but Etcher detected potential corruption issues when reading the image back from the drive. \n\nPlease consider writing the image to a different drive.',
+				'The write has been completed successfully but Spark detected potential corruption issues when reading the image back from the drive. \n\nPlease consider writing the image to a different drive.',
 			openError:
 				'Something went wrong while opening {{source}}.\n\nError: {{error}}',
 			flashError: 'Something went wrong while writing {{image}} {{targets}}.',
 			unplug:
-				"Looks like Etcher lost access to the drive. Did it get unplugged accidentally?\n\nSometimes this error is caused by faulty readers that don't provide stable access to the drive.",
+				"Looks like Spark lost access to the drive. Did it get unplugged accidentally?\n\nSometimes this error is caused by faulty readers that don't provide stable access to the drive.",
 			cannotWrite:
-				'Looks like Etcher is not able to write to this location of the drive. This error is usually caused by a faulty drive, reader, or port. \n\nPlease try again with another drive, reader, or port.',
+				'Looks like Spark is not able to write to this location of the drive. This error is usually caused by a faulty drive, reader, or port. \n\nPlease try again with another drive, reader, or port.',
 			childWriterDied:
-				'The writer process ended unexpectedly. Please try again, and contact the Etcher team if the problem persists.',
+				'The writer process ended unexpectedly. Please try again, and contact the Spark team if the problem persists.',
 			badProtocol: 'Only http:// and https:// URLs are supported.',
 		},
 		target: {
@@ -104,7 +104,7 @@ const translation = {
 			showHidden: 'Show {{num}} hidden',
 			systemDriveDanger:
 				'Selecting your system drive is dangerous and will erase your drive!',
-			openInBrowser: '`Etcher will open {{link}} in your browser`',
+			openInBrowser: '`Spark will open {{link}} in your browser`',
 			changeTarget: 'Change target',
 			largeDriveWarning: 'You are about to erase an unusually large drive',
 			largeDriveWarningMsg:
@@ -150,14 +150,14 @@ const translation = {
 			devTool: 'Toggle Developer Tools',
 			window: 'Window',
 			help: 'Help',
-			pro: 'Etcher Pro',
-			website: 'Etcher Website',
+			pro: 'Spark Pro',
+			website: 'Spark Website',
 			issue: 'Report an issue',
-			about: 'About Etcher',
-			hide: 'Hide Etcher',
+			about: 'About Spark',
+			hide: 'Hide Spark',
 			hideOthers: 'Hide Others',
 			unhide: 'Unhide All',
-			quit: 'Quit Etcher',
+			quit: 'Quit Spark',
 		},
 	},
 };