+++ lib/gui/app/modules/image-writer.ts
@@ -103,7 +103,7 @@ async function performWrite(
 					errors.createUserError({
 						title: 'The writer process ended unexpectedly',
 						description:
-							'Please try again, and contact the Etcher team if the problem persists',
+							'Please try again, and contact the Spark team if the problem persists',
 					}),
 				);
 			}