+++ lib/gui/etcher.ts
@@ -178,16 +178,16 @@ electron.app.on('before-quit', () => {
 
 // this is replaced at build-time with the path to helper binary,
 // relative to the app resources directory.
-declare const ETCHER_UTIL_BIN_PATH: string;
+declare const SPARK_UTIL_BIN_PATH: string;
 
 electron.ipcMain.handle('get-util-path', () => {
 	if (process.env.NODE_ENV === 'development') {
 		// In development there is no "app bundle" and we're working directly with
 		// artifacts from the "out" directory, where this value point to.
-		return ETCHER_UTIL_BIN_PATH;
+		return SPARK_UTIL_BIN_PATH;
 	}
 	// In any other case, resolve the helper relative to resources path.
-	return path.resolve(process.resourcesPath, ETCHER_UTIL_BIN_PATH);
+	return path.resolve(process.resourcesPath, SPARK_UTIL_BIN_PATH);
 });
 
 async function main(): Promise<void> {